Refactor main content script constants
