+++ src/content-scripts/components/viewer/report-editor/Report.tsx
@@ -4,7 +4,7 @@ import classNames from 'classnames';
 import styles from './ReportEditor.scss';
 import { AnnotationAPIModel } from 'content-scripts/api/annotations';
 import { AnnotationReportAPIModel, DataResponse, Reasons } from 'content-scripts/api/annotation-reports';
-import { PPScopeClass } from 'content-scripts/class-consts';
+import { PPScopeClass } from 'content-scripts/settings';
 import ppGA from 'common/pp-ga';
 
 interface IReportProps {
