+++ src/content-scripts/main.tsx
@@ -65,9 +65,3 @@ if (isBrowser) {
     );
   });
 }
-
-// The node within which annotations are made
-// It's lazy so operations on DOM can be done here if needed
-export function annotationRootNode() {
-  return document.body;
-}
