+++ src/content-scripts/utils/TextSelector.ts
@@ -1,10 +1,10 @@
 import { Range as XPathRange } from 'xpath-range';
 import $ from 'jquery';
 import _isEqual from 'lodash/isEqual';
-import { PPHighlightClass } from 'content-scripts/class-consts';
+import { PPHighlightClass } from 'content-scripts/settings';
 // More on xpath-range here: https://github.com/opengovfoundation/xpath-range
 // Wondering what's inside? See https://github.com/opengovfoundation/xpath-range/blob/master/src/range.coffee#L227
-import { annotationRootNode } from '../main';
+import { annotationRootNode } from '../settings';
 
 /**
  * TextSelector monitors a document (or a specific element) for text selections
