+++ src/misc/examples/highlight.ts
